backend/pyproject.toml
[project]
name = "seva-mining-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "apscheduler==3.11.2",
    "twilio>=9.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "anthropic>=0.86.0",
    "pyjwt>=2.9",
    "bcrypt>=4.2,<5",
    "aiosqlite>=0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
